Completed: two processes read, master process gathers

+++ MITgcm_configurations/global_ocean.90x40x15/code/DIAGNOSTICS.h
@@ -194,7 +194,7 @@ C     diagSt_Fname(n)  :: output file name for output stream # n
       LOGICAL   diagSt_ascii, diagSt_mnc
 
       CHARACTER*40 ob_fileName
-      REAL*4 ob_subMask(1-Olx:sNx+Olx,1-Oly:sNy+Oly,nSx,nSy)
+      REAL*8 ob_subMask(1-Olx:sNx+Olx,1-Oly:sNy+Oly,nSx,nSy)
 
       COMMON / DIAG_STATIS_R /
      &     diagSt_freq, diagSt_phase
